backend/requirements.txt
aiofiles
annotated-types
anyio
beautifulsoup4
blinker
Brotli
certifi
cffi
charset-normalizer
click
cryptography
ddgs
decorator
distro
dnspython
duckduckgo-search
ecdsa
elevenlabs
fastapi
gtts
boto3
Flask
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
Flask-Cors
h11
h2
hpack
httpcore
httpx
hyperframe
idna
imageio
imageio-ffmpeg
itsdangerous
Jinja2
lxml
MarkupSafe
moviepy
numpy
openai
anthropic
fal-client
pillow
pip
primp
proglog
pyasn1
pycountry
pycparser
pydantic
pydantic_core
PyJWT
pymongo
python-dotenv
python-jose
python-multipart
pywhatkit
requests
rsa
setuptools
six
sniffio
socksio
soupsieve
starlette
stripe
tqdm
typing_extensions
typing-inspection
urllib3
uvicorn
websockets
Werkzeug
openai-whisper